requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.3.0
autocommand==2.2.2
Automat==24.8.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
bottle==0.13.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cheroot==10.0.1
CherryPy==18.10.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
Django==5.1.4
django-seed==0.3.1
djangorestframework==3.15.2
Faker==33.3.1
filelock==3.16.1
Flask==3.0.3
h11==0.14.0
httpcore==1.0.6
httpie==3.2.4
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
keyboard==0.13.5
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
MouseInfo==0.1.3
multidict==6.1.0
mysqlclient==2.2.7
openai==1.53.0
packaging==24.1
parsel==1.9.1
pillow==10.4.0
pocketsphinx==5.0.3
portend==3.2.0
Protego==0.3.1
psycopg==3.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyOpenSSL==24.3.0
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pyttsx3==2.98
pytweening==1.2.0
pywhatkit==5.4
pywin32==308
queuelib==1.7.0
reportlab==4.2.5
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.11.0
sqlparse==0.5.3
tempora==5.8.0
tldextract==5.1.3
toposort==1.10
tqdm==4.66.6
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
w3lib==2.2.1
Werkzeug==3.0.6
wikipedia==1.4.0
zc.lockfile==3.0.post1
zope.interface==7.2
